core: Match DAs by hw_code as well
For legacy and V6 devices, DAs can be reliably identified using the `hw_code`, resulting in easy matches.

V5 devices, however, require additional matching logic as `hw_code` alone is not sufficient.

This change ensures that when a DA is identifiable via `hw_code`, we attempt to match it before falling back to returning `None`.